requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
